% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta7.R
\name{CV}
\alias{CV}
\title{Converts a function to a package-compatible function}
\usage{
CV(f)
}
\arguments{
\item{f}{A function}
}
\value{
The function bound to the class PDF
}
\description{
Converts a function to a package-compatible function
}
\examples{
a <- CV(f)
b <- CV(g)
}
