% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta9.R
\name{utocdf}
\alias{utocdf}
\title{Uniform 0-1 number to CDF-based number}
\usage{
utocdf(CDF, num)
}
\arguments{
\item{CDF}{A CDF}

\item{num}{A number}
}
\value{
The number \code{num} mapped according to the \code{CDF}
}
\description{
Uniform 0-1 number to CDF-based number
}
\examples{
utocdf(f, 0.3)
uto(f, 0.2)
}
